+++ Gemfile
@@ -1,9 +1,3 @@
 source 'https://rubygems.org'
 
-gem 'rambling-trie-opal', :git => 'https://github.com/interscript/rambling-trie.git'
-gem 'opal', '~> 1.0.3'
-gem 'guard'
-gem 'guard-rake'
-gem "closure-compiler", github: "hmdne/closure-compiler"
-
 gemspec
