+++ lib/interscript-opal.rb
@@ -1,0 +1,2 @@
+require "opal"
+require "interscript"
