--- opal/interscript-opal.rb
@@ -1,2 +1,0 @@
-require "opal"
-require "interscript"
